tts/indic_tts/requirements.txt
# Indic TTS Requirements
# Core TTS library (Coqui TTS - fork with Python 3.12 support)
coqui-tts>=0.27.0

# PyTorch (with CUDA support)
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.0

# API
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6

# Other dependencies
gdown>=4.7.0  # For downloading model weights
